Calcul des angles de camera compris et mis a jour dans camera_move

+++ srcs/print_tab_debug.c
@@ -1,7 +1,7 @@
 #include "../includes/fdf.h"
 #include "../libft/include/libft.h"
 
-void	print_tab_debug(float ***TABTAB)
+void	print_tab_debug(double ***TABTAB)
 {
 	int i;
 	int j;
